+++ Dockerfile
@@ -42,4 +42,4 @@ COPY . .
 EXPOSE 8000
 
 # Run the application.
-CMD ["uvicorn" , "app.main:app", "--host", "0.0.0.0", "--port", "8000"]
+CMD ["uvicorn" , "app:app", "--host", "0.0.0.0", "--port", "8000"]
